feat(cbus): implemented security application

+++ plc4go/protocols/cbus/readwrite/model/LightingDataLabel.go
@@ -204,7 +204,7 @@ func LightingDataLabelParse(readBuffer utils.ReadBuffer, commandTypeContainer Li
 		}
 	}
 	// Byte Array field (data)
-	numberOfBytesdata := int(uint16(commandTypeContainer.NumBytes()) - uint16(uint16(utils.InlineIf(bool(bool((labelOptions.GetLabelType()) != (LightingLabelType_LOAD_DYNAMIC_ICON))), func() interface{} { return uint16(uint16(3)) }, func() interface{} { return uint16(uint16(2)) }).(uint16))))
+	numberOfBytesdata := int(uint16(uint16(commandTypeContainer.NumBytes()) - uint16(uint16(utils.InlineIf(bool(bool((labelOptions.GetLabelType()) != (LightingLabelType_LOAD_DYNAMIC_ICON))), func() interface{} { return uint16(uint16(uint16(3))) }, func() interface{} { return uint16(uint16(uint16(2))) }).(uint16)))))
 	data, _readArrayErr := readBuffer.ReadByteArray("data", numberOfBytesdata)
 	if _readArrayErr != nil {
 		return nil, errors.Wrap(_readArrayErr, "Error parsing 'data' field of LightingDataLabel")
